+++ tests/fixtures/source/MainCompoundAssignment.xs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+fn main() -> Long {
+    value: Long = 2;
+    value += 5;
+    value *= 2;
+    value -= 4;
+    value /= 3;
+    value %= 2;
+    value ^= 6;
+    value &= 7;
+    value |= 0;
+    return value;
+}